# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Test",
        "description": "Test API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://muntian.softwars.com.ua",
            "description": "API"
        }
    ],
    "paths": {
...
        "/api/tasks/batch": {
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Создать несколько задач",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/BatchTasksRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/TasksViewResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/error:bad_request"
                    },
                    "422": {
                        "$ref": "#/components/responses/error:validation"
                    },
                    "500": {
                        "$ref": "#/components/responses/error:server"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/api/tasks/{id}": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Просмотреть задачу",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/TasksViewResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/error:bad_request"
                    },
                    "422": {
                        "$ref": "#/components/responses/error:validation"
                    },
                    "500": {
                        "$ref": "#/components/responses/error:server"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Обновить задачу",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/TasksRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/TasksViewResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/error:bad_request"
                    },
                    "422": {
                        "$ref": "#/components/responses/error:validation"
                    },
                    "500": {
                        "$ref": "#/components/responses/error:server"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Удалить задачу",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "example": "1"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "400": {
                        "$ref": "#/components/responses/error:bad_request"
                    },
                    "422": {
                        "$ref": "#/components/responses/error:validation"
                    },
                    "500": {
                        "$ref": "#/components/responses/error:server"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "BatchTasksRequest": {
                "required": [
                    "data"
                ],
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "required": [
                                "name",
                                "description",
                                "category",
                                "deadline",
                                "urgent",
                                "image"
                            ],
                            "properties": {
                                "name": {
                                    "description": "Class BatchRequest",
                                    "type": "string",
                                    "example": "test"
                                },
                                "description": {
                                    "type": "string",
                                    "example": "test"
                                },
                                "category": {
                                    "type": "string",
                                    "enum": [
                                        "work",
                                        "personal"
                                    ],
                                    "example": "work"
                                },
                                "deadline": {
                                    "type": "string",
                                    "format": "date",
                                    "example": "2022-10-05"
                                },
                                "urgent": {
                                    "type": "boolean",
                                    "example": "false"
                                },
                                "image": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "TasksRequest": {
                "required": [
                    "name",
                    "description",
                    "category",
                    "deadline",
                    "urgent",
                    "image"
                ],
                "properties": {
                    "name": {
                        "description": "Class Request",
                        "type": "string",
                        "example": "test"
                    },
                    "description": {
                        "type": "string",
                        "example": "test"
                    },
                    "category": {
                        "type": "string",
                        "enum": [
                            "work",
                            "personal"
                        ],
                        "example": "work"
                    },
                    "deadline": {
                        "type": "string",
                        "format": "date",
                        "example": "2022-10-05"
                    },
                    "urgent": {
                        "type": "boolean",
                        "example": "false"
                    },
                    "image": {
                        "type": "string",
                        "format": "binary"
                    }
                },
                "type": "object"
            },
...
            "TasksViewResource": {
                "properties": {
                    "id": {
                        "description": "Class ViewResource",
                        "type": "integer",
                        "example": "234"
                    },
                    "name": {
                        "type": "string",
                        "example": "test"
                    },
                    "description": {
                        "type": "string",
                        "example": "test"
                    },
                    "category": {
                        "type": "string",
                        "enum": [
                            "work",
                            "personal"
                        ],
                        "example": "work"
                    },
                    "deadline": {
                        "type": "string",
                        "format": "date",
                        "example": "2022-10-05"
                    },
                    "completed": {
                        "type": "boolean",
                        "example": "false"
                    },
                    "urgent": {
                        "type": "boolean",
                        "example": "false"
                    },
                    "image": {
                        "type": "string",
                        "format": "uri",
                        "example": "http://test.dok/storage/52/download.jpeg"
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
...
            "error:bad_request": {
                "description": "Bad request",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "description": "Текст ошибки",
                                    "type": "string",
                                    "example": "messages.exception.bad_request"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
            "error:server": {
                "description": "Internal server error",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "description": "Текст ошибки",
                                    "type": "string",
                                    "example": "messages.exception.server"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
            "error:validation": {
                "description": "Data validation failed",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "messages.exception.given_data_invalid"
                                },
                                "errors": {
                                    "properties": {
                                        "field": {
                                            "type": "string",
                                            "example": "Field is required"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "Токен авторизации, полученный при логине.<b>Authorization: Bearer {token}</b>",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "bearer": []
        }
    ]
}